By using closures instead of lambda / caching field parsers methods / precompiling regular objects, a performance improvement of about 15% was achieved (mainly from closures)
```bash
(pystdf) ➜  pystdf git:(speed_up) ✗ python test.py                                             
Time taken to import data/lot3.stdf: 0.4714 seconds for 4.35 MB of data (lot3)
Time taken to import data/lot2.stdf: 0.4827 seconds for 4.21 MB of data (lot2)
Time taken to import data/demofile.stdf: 0.4606 seconds for 4.35 MB of data (demofile)
(pystdf) ➜  pystdf git:(speed_up) ✗ git switch master
(pystdf) ➜  pystdf git:(master) ✗ python test.py
Time taken to import data/lot3.stdf: 0.5451 seconds for 4.35 MB of data (lot3)
Time taken to import data/lot2.stdf: 0.5531 seconds for 4.21 MB of data (lot2)
Time taken to import data/demofile.stdf: 0.5512 seconds for 4.35 MB of data (demofile)
```

Here are my benchmark codes:
```python3
from pystdf.Importer import ImportSTDF
from time import time
from pathlib import Path


if __name__ == "__main__":
    root = Path(".")
    for std_file in (root / "data").glob("*.stdf"):
        start_time = time()
        data = ImportSTDF(std_file)
        print(
            f"Time taken to import {std_file}: {(time() - start_time):.4f} seconds for {std_file.stat().st_size / 1024**2:.2f} MB of data ({std_file.stem})"
        )
```