@cmars 2025-04-08
Awesome, thanks for this! Will try to get to this later this evening.