@Asachoo 2025-04-09
From a performance standpoint, batch processing of `readCn` seems like a promising next step for performance improvements.

However, this PR is already quite complex, and introducing `batchReadFields` has made it even more difficult to review and maintain.

To keep things manageable, I think it would be better to address the optimization of `readCn` in a separate PR.

This way, we can focus on refining the current changes while ensuring clarity and maintainability.

![img_v3_02l6_9a4fae9f-90aa-4efc-8493-58a0e6cf1f8g](https://github.com/user-attachments/assets/1f7852cb-49c4-415b-8ad5-1d87e1dd50b4)

Let me know your thoughts!